Disable check for declarations not being separated by empty lines
Because this check applies to abstract method declarations inside
interfaces (which we may very well want to write in compact form, as was
done in previous ITProFound web apps), and there's no clear way to
selectively disable the check in this circumstance, we'll disable it
everywhere.  We retain the check prohibiting multiple blank lines.