Implement superadmin user type and automatic registration (#20)
* Implement superadmin user type and automatic registration

When the application is started, a single superadmin user is
automatically created and stored in the database with the other users.
The username and password for this user must be supplied via Spring
properties (e.g., in the `application.properties` file or through
environment variables); if either is null, no superadmin will be
registered.

If the database already contains a user with the specified username, a
new user is not created, and instead this user's password is updated to
match the supplied property.  This can be used to update the
superadmin's password, though other mechanisms that do not involve
restarting the application may be considered in the future.  Changing
an existing superadmin's username this way is not recommended, because
no attempt is made to discover and delete superadmin users with other
usernames.

* Remove unnecessary `@PropertySource` annotation

The `application.properties` file is automatically picked up by Spring
after all.  I originally thought it wasn't, but that was only because
Maven was failing to copy the file into the build folder, and I failed
to test that the annotation was necessary after I fixed the issue with
Maven.

* Remove unnecessary block in the pom concerning `application.properties`

It's still unclear what I was doing wrong before, but Maven does seem to
be picking up the file appropriately without this now.