Implement Basic Validation for Customer Creation (#25)
* Refactor controller tests

* Implement basic validation for customer creation

This validation is currently very crude, in that it returns only an
error code in the event of validation failure.  Once we have a view,
we'll need to move to a more robust version that also sends the actual
validation errors to the client; fortunately, this should be relatively
easy to achieve (we just need to change the body type of
`ResponseEntity`, probably to some variant of `Map`).