Implement validation for ODL offices (#29)
* Implement very basic validation for office creation

* Restructure `JsonDto` to reduce duplication of validation test logic

After adding validation code to `OfficeDto`, the tests in
`OfficeControllerTests` naturally echo a good deal of logic already
present in `CustomerControllerTests`.  This is an effort to move that
logic into a single location, with most of the benefit being realized by
specializing `JsonDto` to only accept payloads of type
`Map<String, Serialization>`; this allows us to move methods
manipulating `JsonDto` objects from the test classes into the `JsonDto`
class itself.

* Add basic "alphabetic-only" validation for ODL office names

* Reorganize resource-creation controller tests again

* Encapsulate alphabetic constraint as annotation

* Add validation for address lines (not including city, state, etc.)

* Add validation for city name, and weaken `@Alphabetic` restrictions

In particular, add support for specific punctuation marks (periods,
apostrophes, hyphens, and so forth) to elements validated by the former
`@Alphabetic` annotation (which isn't really alphabetic anymore, so we
rename it to `@Name`).  Since we are also apparently expected to support
cities with numbers in their names, we extend this with a separate
`@AlphanumericName` annotation.

Also, shorten the method names `assertResourcesWithAttributeValuesAre*`
to `assertAttributeValues*`; these are used very frequently throughout
`CreateResourceBaseTests` and its subclasses, and the shorter name
improves readability and code style.

* Add validation for office and city codes

* Add validation for postal code

* Add validation for `state` and `country` fields

* Add phone number validation

* Implement email validation for ODL offices

* Allow numbers in ODL office names

This is per the discussion at today's meeting.

* Shorten resource validation assertion method names