Implement Office Admin creation (#31)
* Extract common data for most users from `Customer` into new base class

This will prevent duplication of code as we proceed to add a new class
representing the ODL Office Admin user type.

Note that this change results in the `Customer` type being slightly
denormalized, since the email is now stored twice, once as itself and
once as the username.

* Make Hibernate entities non-final

This is recommended in the Hibernate docs, as it permits lazy-loading of
associated entities.

* Create `OfficeAdmin` domain model class

* Rename `*ControllerTests` to `*CreationTests`

The idea is that these test classes are testing only the creation method
(the "C" in CRUD) in their associated controller, and other methods will
get their own test classes.  In particular, this will allow the
`SuperadminController` class to support CRUD operations for both
`Office`s and `OfficeAdmin`s without running into trouble regarding the
type parameter on `JsonDtoValidationTester`.

This may not be the best solution; maybe the controller tests should
access `JsonDtoValidationTester` through composition instead of
inheritance.  On the other hand, that change would require significantly
more work, and it's not clear to me that the result would be cleaner or
simpler than just having multiple test classes.

* Implement basic creation method for ODL Office Admins

No validation is currently performed; that will come next.

* Implement validation for most Office Admin fields

The one exception is the `username` field.  Since this will eventually
be removed from the form and automatically generated, it doesn't make
much sense to spend effort validating it now.