Implement automatic generation of usernames (#32)
* Add mechanism for counting number of users per user type

This will be used in the machinery for automatic generation of
usernames.

* Implement automatic generation of usernames for Office Admins

This also establishes the general design for automatic username
generation. `User` subtypes that wish to implement automatic username
generation just need to implement the `User.AutogeneratesUsername`
interface.

* Remove explicit username from Office Admin creation process

* Add transactional control to automatic username generation

This is necessary because our name generation scheme involves the
creation of a unique index number, and the implementation we have chosen
for getting that index number requires a preliminary query to the
database, thereby introducing an obvious race condition.  (It might be
possible, or even trivial, to replace the current scheme with one using
in-database sequences, in which case explicit transaction control would
be unnecessary, but how to go about doing this isn't particularly
obvious, at least to me.)